requirements.txt
bs4==0.0.2
cachetools==5.3.2
chromadb== 0.4.14
dataclasses-json==0.6.3
duckduckgo_search==4.2
fastapi==0.109.0
huggingface-hub==0.20.2

jsonpatch==1.33
jsonpath-python==1.0.6
langchain==0.1.1
langchain-community==0.0.13
langchain-core==0.1.12
langchain_openai
langsmith==0.0.83


nltk==3.8.1
numpy

openai==1.8.0
ollama==0.1.6
pandas
pdf2image==1.17.0

pydantic==2.5.3
pydantic_core==2.14.6

pypdf==3.17.4
PyPDF2==3.0.1
pypdfium2==4.26.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0

requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
safetensors==0.4.1

scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99

SQLAlchemy==2.0.25
starlette==0.35.1
streamlit==1.30.0

threadpoolctl==3.2.0
tiktoken==0.5.2
timm==0.9.12
tokenizers==0.15.0


unstructured==0.12.0
unstructured-client==0.15.2
unstructured-inference==0.7.21
unstructured.pytesseract==0.3.12
uvicorn==0.26.0
wikipedia==1.4.0

